pyproject.toml
[build-system]
requires = ["uv_build>=0.8.22"]
build-backend = "uv_build"

[project]
name = "aind-behavior-iso-force"
description = "A library that defines AIND data schema for the Aind Behavior IsoForce experiment."
authors = [
    { name = "Bruno Cruz", email = "bruno.cruz@alleninstitute.org" },
]
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
version = "0.1.0rc0"
readme = {file = "README.md", content-type = "text/markdown"}

dependencies = [
  "aind_behavior_services>=0.13.2",
  "pydantic-settings"
]

[project.urls]
Documentation = "https://allenneuraldynamics.github.io/Aind.Behavior.IsoForce/"
Repository = "https://github.com/AllenNeuralDynamics/Aind.Behavior.IsoForce/"
Issues = "https://github.com/AllenNeuralDynamics/Aind.Behavior.IsoForce/issues"
Changelog = "https://github.com/AllenNeuralDynamics/Aind.Behavior.IsoForce/releases"

[project.optional-dependencies]

launcher = ["aind-clabe[aind-services] >= 0.10"]

data = ["contraqctor>=0.5.3"]

[dependency-groups]

dev = [
    "aind_behavior_iso_force[launcher]",
    "aind_behavior_iso_force[data]",
    'ruff',
    'codespell'
]

docs = [
    'Sphinx',
    'furo',
    'sphinx-jinja',
    'autodoc_pydantic[erdantic]',
    'myst_parser',
    'sphinx-jsonschema',
    "aind_behavior_iso_force[data]",  # Necessary for the data contract rendering
]

[project.scripts]
iso-force = "aind_behavior_iso_force.cli:main"


[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.ruff.lint]
extend-select = ['Q', 'RUF100', 'C90', 'I']
extend-ignore = []
mccabe = { max-complexity = 14 }
pydocstyle = { convention = 'google' }

[tool.codespell]
skip = '.git,*.pdf,*.svg,./bonsai,*.bonsai,./docs/_build,uv.lock,*.Generated.cs'
ignore-words-list = 'nd'

[tool.uv]
default-groups = ['dev']
required-version = '>=0.8.4'
